requirements.txt
mcp==1.6.0
anthropic==0.49.0
python-dotenv==1.1.0
streamlit==1.44.1fastapi==0.115.12
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
uvicorn==0.34.0
